refactor(navbar): remove Navbar.Layout in favor of just Navbar
BREAKING CHANGE: Navbar.Layout is now just Navbar
