+++ src/components/organisms/Navbar/index.md
@@ -1,6 +1,6 @@
 ### Summary
 
-`<Navbar />` doesn't render anything: it's just a namespace for `<Navbar.Dropdwn />`, `<Navbar.Dropdown />` and `<Navbar.Link />`.
+`<Navbar />` is a wrapper and a namespace for `<Navbar.Dropdown />` and `<Navbar.Link />`.
 
 In order to create a navigation bar to be used within a **Zopa application**, you will need to leverage these components together.
 
@@ -11,7 +11,7 @@ In order to create a navigation bar to be used within a **Zopa application**, yo
 ```js { "props": { "style": { "transform": "translate3d(0, 0, 0)", "border": "2px solid #efefef" } } }
 import { Navbar, Link, HamburgerIcon, ProfileIcon } from '@zopauk/react-components';
 
-<Navbar.Layout
+<Navbar
   backgroundColor="#00B9A7"
   left={
     <Navbar.Dropdown
@@ -102,7 +102,7 @@ class ResponsiveNavbar extends React.Component {
   }
 
   render() {
-    return <Navbar.Layout backgroundColor="#00B9A7" left={this.renderLeft()} />;
+    return <Navbar backgroundColor="#00B9A7" left={this.renderLeft()} />;
   }
 }
